% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{doiref}
\alias{doiref}
\title{Fetch Bibliographies via DOI}
\usage{
doiref(path, style = "apa")
}
\arguments{
\item{path}{file path}

\item{style}{citaiton style, default apa}
}
\description{
Fetch Bibliographies via DOI
}
