% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlescholar.R
\name{getgsauthor}
\alias{getgsauthor}
\title{Fetch author's publication from Google Scholar}
\usage{
getgsauthor(id, start = 1, end = 100)
}
\arguments{
\item{id}{the google scholar ID to be searched}

\item{start}{begin item}

\item{end}{finish item}
}
\value{
tibble object
}
\description{
Fetch author's publication from Google Scholar
}
\examples{
\dontrun{
library(scifetch)
id <- "SPNX8oUAAAAJ"
z <- getgsauthor(id)
}
}
