% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed.R
\name{getpubmedtbl}
\alias{getpubmedtbl}
\title{Get the basic data from xml}
\usage{
getpubmedtbl(xml2)
}
\arguments{
\item{xml2}{the xml object from getpubmed}
}
\value{
tibble object
}
\description{
Get the basic data from xml
}
\examples{
\dontrun{
library(scifetch)
query <- 'janusz pawliszyn[AU]'
z <- getpubmed(query)
z2 <- getpubmedtbl(z)}
}
\seealso{
getpubmed
}
