% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss.R
\name{getrss}
\alias{getrss}
\title{Fetch rss into tibble}
\usage{
getrss(feed, abslength = 500)
}
\arguments{
\item{feed}{feed address}

\item{abslength}{lenghth of the description, default 500}
}
\value{
tibble object
}
\description{
Fetch rss into tibble
}
\examples{
\dontrun{
library(scifetch)
feed <- 'yihui.name/index.xml'
z <- getrss(feed)
}
}
\seealso{
getpubmed
}
